Remove previous code
