--- analogue.h
@@ -1,10 +1,0 @@
-#ifndef __ANALOGUE_H
-#define __ANALOGUE_H
-
-#include <stdint.h>
-
-extern uint32_t a, b, c, d;
-
-void analogue_init(void);
-
-#endif /* __ANALOGUE_H */
