+++ bootloader/force_bootloader.c
@@ -1,7 +1,5 @@
 #include <stdbool.h>
 
-#include "button.h"
-#include "smps.h"
 #include <libopencm3/stm32/gpio.h>
 #include <libopencm3/stm32/rcc.h>
 
@@ -10,21 +8,20 @@
 #define EXT_PORT GPIOC
 #define EXT_PIN GPIO15
 
+#define EN_PORT GPIOB
+#define EN_PIN GPIO5
+#define TRIM_PORT GPIOC
+#define TRIM_PIN GPIO12
+
 #define delay(x) do { for (int i = 0; i < x * 1000; i++) \
                           __asm__("nop"); \
                     } while(0)
-
-void button_init(void) {
-	gpio_set(INT_PORT, INT_PIN); // Pull-up
-	gpio_set_mode(INT_PORT, GPIO_MODE_INPUT, GPIO_CNF_INPUT_PULL_UPDOWN, INT_PIN);
-	gpio_set(EXT_PORT, EXT_PIN); // Pull-up
-	gpio_set_mode(EXT_PORT, GPIO_MODE_INPUT, GPIO_CNF_INPUT_PULL_UPDOWN, EXT_PIN);
-}
 
 bool button_int_read(void) { return gpio_get(INT_PORT, INT_PIN); }
 bool button_ext_read(void) { return gpio_get(EXT_PORT, EXT_PIN); }
 
 
+void smps_on_boot(void) __attribute__((section(".bootloader")));
 bool force_bootloader() __attribute__((section(".bootloader")));
 
 bool force_bootloader()
@@ -55,7 +52,14 @@ bool force_bootloader()
 	return false;
 }
 
-bool button_pressed(void)
-{
-	return !button_int_read() || !button_ext_read();
+void smps_on_boot(void) {
+	// Turn SMPS on in the context of the bootloader
+	rcc_periph_clock_enable(RCC_GPIOB);
+	rcc_periph_clock_enable(RCC_GPIOC);
+
+	gpio_set_mode(EN_PORT, GPIO_MODE_OUTPUT_2_MHZ, GPIO_CNF_OUTPUT_PUSHPULL, EN_PIN);
+	gpio_set_mode(TRIM_PORT, GPIO_MODE_OUTPUT_2_MHZ, GPIO_CNF_OUTPUT_OPENDRAIN, TRIM_PIN);
+
+	gpio_set(EN_PORT, EN_PIN);
+	gpio_set(TRIM_PORT, TRIM_PIN);
 }
