--- button.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include <stdbool.h>
-
-void button_init(void);
-
-bool button_pressed(void);
