--- fan.h
@@ -1,7 +1,0 @@
-#pragma once
-
-void fan_init(void);
-
-void fan_on(void);
-
-void fan_off(void);
