--- oocd/pbv4.cfg
@@ -1,11 +1,0 @@
-# Hitex stm32 performance stick
-
-add_script_search_dir /usr/local/share/openocd/scripts
-
-reset_config srst_only
-
-source [find interface/ftdi/olimex-arm-usb-ocd-h.cfg]
-
-set  CHIPNAME stm32_pbv4
-set  WORKAREASIZE 0x1000
-source [find target/stm32f1x.cfg]
