--- pswitch.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include <stdbool.h>
-
-void pswitch_init(void);
-
-bool pswitch_read(void);
