--- smps.h
@@ -1,12 +1,0 @@
-#pragma once
-
-void smps_init(void);
-
-void smps_on(void);
-void smps_on_boot(void);
-
-void smps_off(void);
-
-void smps_boost_on(void);
-
-void smps_boost_off(void);
