--- stm32-pbv4_test.ld
@@ -1,7 +1,0 @@
-MEMORY
-{
-        rom (rx) : ORIGIN = 0x08000000, LENGTH = 32K
-        ram (rwx) : ORIGIN = 0x20000000, LENGTH = 10K
-}
-
-INCLUDE libopencm3_stm32f1.ld
