--- temp.c
@@ -1,5 +1,0 @@
-#include "temp.h"
-
-void temp_init(void) {
-
-}
