--- temp.h
@@ -1,3 +1,0 @@
-#pragma once
-
-void temp_init(void);
