--- usart.h
@@ -1,8 +1,0 @@
-#ifndef __USART_H
-#define __USART_H
-
-void usart_init(void);
-
-int usart_get_char(void);
-
-#endif /* __USART_H */
