--- usb.h
@@ -1,26 +1,0 @@
-#define POWERBOARD_WRITE_OUTPUT0	0
-#define POWERBOARD_WRITE_OUTPUT1	1
-#define POWERBOARD_WRITE_OUTPUT2	2
-#define POWERBOARD_WRITE_OUTPUT3	3
-#define POWERBOARD_WRITE_OUTPUT4	4
-#define POWERBOARD_WRITE_OUTPUT5	5
-#define POWERBOARD_WRITE_RUNLED		6
-#define POWERBOARD_WRITE_ERRORLED	7
-#define POWERBOARD_WRITE_PIEZO		8
-
-#define POWERBOARD_READ_OUTPUT0		0
-#define POWERBOARD_READ_OUTPUT1		1
-#define POWERBOARD_READ_OUTPUT2		2
-#define POWERBOARD_READ_OUTPUT3		3
-#define POWERBOARD_READ_OUTPUT4		4
-#define POWERBOARD_READ_OUTPUT5		5
-#define POWERBOARD_READ_5VRAIL		6
-#define POWERBOARD_READ_BATT		7
-#define POWERBOARD_READ_BUTTON		8
-#define POWERBOARD_READ_FWVER		9
-
-void usb_init();
-void usb_deinit();
-void usb_poll();
-
-extern bool re_enter_bootloader;
